+++ CHANGELOG.md
@@ -3,6 +3,10 @@
 
 This is a *minor release* that aims to be fully compatible with v0.2.0 while fixing bugs.
 
+List of bugs fixed:
+* Maximum memory setting is sometimes ignored (https://github.com/qupath/qupath/issues/582)
+  * Memory cannot no longer be specified to be less than 1 GB
+
 
 ## Version 0.2.2